PO-2654/ Refactor dto objects to correct directory
