+++ src/main/java/uk/gov/hmcts/opal/dto/history/AccountHistoryDetails.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.opal.dto.history;
+
+public interface AccountHistoryDetails {
+}
