+++ src/main/java/uk/gov/hmcts/opal/dto/history/AccountHistoryEnforcementDetails.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.service.opal.history.core;
+package uk.gov.hmcts.opal.dto.history;
 
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
