--- src/main/java/uk/gov/hmcts/opal/service/opal/history/core/AccountHistoryDetails.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.opal.service.opal.history.core;
-
-public interface AccountHistoryDetails {
-}
