+++ src/main/java/uk/gov/hmcts/opal/service/opal/history/defendant/DefendantAccountHistoryService.java
@@ -10,10 +10,10 @@ import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryResponse;
 import uk.gov.hmcts.opal.service.opal.history.HistoryItemOrderingService;
 import uk.gov.hmcts.opal.service.opal.history.core.AbstractAccountHistoryService;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryContext;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryItem;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryResult;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryType;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryContext;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryItem;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryResult;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryType;
 import uk.gov.hmcts.opal.service.opal.history.defendant.sources.AmendmentHistorySource;
 import uk.gov.hmcts.opal.service.opal.history.defendant.sources.DefendantTransactionHistorySource;
 import uk.gov.hmcts.opal.service.opal.history.defendant.sources.EnforcementHistorySource;
