+++ src/test/java/uk/gov/hmcts/opal/service/opal/history/defendant/DefendantAccountHistoryModelAdapterTest.java
@@ -13,12 +13,12 @@ import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryItem;
 import uk.gov.hmcts.opal.dto.history.HistoryItemType;
 import uk.gov.hmcts.opal.dto.history.NoteDetails;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryDetails;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryFilter;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryItem;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryItemType;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryNoteDetails;
-import uk.gov.hmcts.opal.service.opal.history.core.AccountHistoryPostedDetails;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryDetails;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryFilter;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryItem;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryItemType;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryNoteDetails;
+import uk.gov.hmcts.opal.dto.history.AccountHistoryPostedDetails;
 
 class DefendantAccountHistoryModelAdapterTest {
